requirements.txt
langchain-mcp-adapters
mcp
langgraph
langchain
langchain-google-genai
python-dotenv
langchain_community
youtube-transcript-api
datetime
os
re
httpx
asyncio